requirements.txt
pandas 
requests
streamlit 
matplotlib.pyplot 
datetime 